feat: solve kata Sum of pairs
The essence of the algorithm is to find a specific number,
because we know the sum and one of the numbers of this
sum. This means that it is necessary to go through the
array and as soon as the desired number is encountered
(`s - i = x`, where `s`, `i` are known, `x` is the desired number)
return a pair of numbers.

Refs: https://www.codewars.com/kata/54d81488b981293527000c8f/rust