feat!: Split the project into separate packages
The implementation of the CLI for working with TMDB has been moved
to a separate package in order to be able to assemble the project
in separate executable files.

To run the tmdb_cli package, you need to run
`cargo run -p tmdb_cli` specifying a specific package using the `-p`
attribute (`--package <package>`).

Tasks with codewars remained in the main package, which can be run
by running the `cargo test'.